SensorManager/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
ipython==8.12.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kafka-python==2.0.2
kazoo==2.9.0
numpy==1.24.2
oauthlib==3.2.2
packaging==23.1
pandas==1.5.3
pipreqs==0.4.11
portalocker==2.7.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2-binary==2.9.5
PyJWT==2.6.0
requests==2.28.2
requests-oauthlib==1.3.1
Resource==0.2.1
six==1.16.0
SQLAlchemy==2.0.9
Werkzeug==2.2.3
